requirements.txt
# Core Deep Learning Libraries
torch>=1.10.0
torchvision
numpy>=1.20.0
matplotlib>=3.3.4
Pillow>=8.0.0
scikit-learn>=1.0.0

# Deployment and API
fastapi>=0.70.0  # Used for the face_recognition_api.py
uvicorn[standard]>=0.15.0 # ASGI server to run FastAPI
python-multipart # Used by FastAPI to handle form data (like file uploads)

# Optional (If used for data handling/plotting in notebook)
pandas>=1.2.0